renamed homepage and added images folder
